+++ enterprise/internal/campaigns/store_test.go
@@ -1645,8 +1645,14 @@ func testStorePatchSets(t *testing.T, ctx context.Context, s *Store, _ repos.Sto
 	})
 }
 
-func testStorePatches(t *testing.T, ctx context.Context, s *Store, _ repos.Store, clock clock) {
+func testStorePatches(t *testing.T, ctx context.Context, s *Store, reposStore repos.Store, clock clock) {
 	patches := make([]*cmpgn.Patch, 0, 3)
+
+	repo := testRepo(1, "github")
+	deletedRepo := testRepo(2, "github").With(repos.Opt.RepoDeletedAt(clock.now()))
+	if err := reposStore.UpsertRepos(ctx, deletedRepo, repo); err != nil {
+		t.Fatal(err)
+	}
 
 	var (
 		added   int32 = 77
@@ -1658,7 +1664,7 @@ func testStorePatches(t *testing.T, ctx context.Context, s *Store, _ repos.Store
 		for i := 0; i < cap(patches); i++ {
 			p := &cmpgn.Patch{
 				PatchSetID: int64(i + 1),
-				RepoID:     1,
+				RepoID:     repo.ID,
 				Rev:        api.CommitID("deadbeef"),
 				BaseRef:    "master",
 				Diff:       "+ foobar - barfoo",
@@ -1695,6 +1701,21 @@ func testStorePatches(t *testing.T, ctx context.Context, s *Store, _ repos.Store
 			patches = append(patches, p)
 		}
 	})
+
+	// Create patch to deleted repo.
+	{
+		p := &cmpgn.Patch{
+			PatchSetID: 1000,
+			RepoID:     deletedRepo.ID,
+			Rev:        api.CommitID("deadbeef"),
+			BaseRef:    "master",
+			Diff:       "+ foobar - barfoo",
+		}
+		err := s.CreatePatch(ctx, p)
+		if err != nil {
+			t.Fatal(err)
+		}
+	}
 
 	t.Run("Count", func(t *testing.T) {
 		count, err := s.CountPatches(ctx, CountPatchesOpts{})